refactor: move Subject List Table to own comp
